proc: bugfix: Next, StepOut: normal calls to deferred functions
When a deferred function is called directly Next and StepOut should not
step into it if it is called normally.

This commit alters previous behaviour of Next and StepOut. Previously
Next (and StepOut) would step into a deferred function called while
returning from the current function, with this commit they will only
step into a deferred function if it is called during panic.

The reason for this is twofold

1. It is much harder to distinguish deferreturn calls from normal calls
than it is distinguishing panic calls from normal calls. It is still
possible but would involve disassembling the caller function

2. It is not the correct behavior, requesting next on the last line of
a function, or stepout, should not result in stepping into a call to a
deferred function.