Add a note about headless and init
I spent some time trying to run an init file in headless mode and I just
discover that this feature is not implemented yet.

So I think it is reasonable to add this note to make it more clear for
other people that are trying to do the same.

In second, I would like to have this feature even in headless mode so, I
will open an issue to track it down.

Signed-off-by: Gianluca Arbezzano <gianarb92@gmail.com>