Add capability to jump to location and skip execution of instructions
This currently is a more or less work in progress, it depends on the
feature set we want to support this feature. In my opinion, this could
be flagged as an experimental feature and if the user has an issue then
it's their fault. Maybe it's not a 100% good statement to have. There
are cases where this could be useful to have and a user that understands
how to use it, will be able to use it to the full extent.

I've tested this only on Windows, I'll let the integration test do the
job on the other systems.

Cross function jumping works as well, but only in the toy example below,
I have no clue how the runtime will be affected by this.

Fixes #1045