function backtraces of linux/arm64 core dumps are now printed correctly
Backtracing was greatly improved on linux/arm64 core-file backtraces.
Previously, the backtrace of the goroutine that faulted and caused a core
dump was quite truncated.  Now, this goroutine's backtrace is correctly
printed out.

Faults cause exceptions.  Exceptions are taken on a different stack from
the goroutine which caused the exception.  Further, the exception's stack
frame is defined by the operating system (linux in this case) and not by
golang's runtime environment.  Thus, to correctly navigate a backtrace,
the stack unwinding code must have knowledge of how to process such exception
stack frames.  On the amd64 architecture, golang maintains explicit frame
pointers, easing this task.  On the arm64 architecture, frame pointers are
implicit, complicating this task.

Following gdb's lead in processing exception frames, code was added for the
linux/arm64 implementation to discover that a stack frame is an exception
frame.  If it discovers an exception stack frame, it navigates to the stack
frame that was interrupted.  This code is in the new source file
"arm64_linux_stack.go".

The procedure calls are now correctly printed out.  However, if the go code
was compiled with code optimization on, local variables are often not printed
out correctly.  This has not yet been investigated.  A tentative working
hypothesis is that the local-variable backtrace-following code also needs
to notice the exception record, and if one is detected, fetch the requested
register out of the exception record's register save area.

For now, the code has implicit knowledge of how to advance over an exception
stack frame by having explicit constant offsets.  A better way would be to
look up local stack symbols go's exception handler's code to fetch the
correct stack pointer and/or offset.  However, this will need to wait until
the fetching of local variables out of the stack frames is improved.

Finally, about 12% of the time the backtrace of a core-file was still
truncated.  This turned out to because of original:

  for i := range framectx.Regs {

loop in "stack.go".  As "framectx.Regs" is a map, go can return indices
in a pseudo-random order.  If the record for key "31" is returned before
the record for key "30", something in Delve/Dwarf's register-fix-up code
is done incorrectly.  The code was changed to copy the key/value pairs
into a slice.  The slice was then sorted, the for loop was changed to range
over the slice, and the bug went away.